boa-workshop/pyproject.toml
[project]
name = "boa-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "titanoboa>=0.2.5",
]
